move Product related products feature from project-base to packages
