+++ project-base/app/src/Form/Admin/ProductFormTypeExtension.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace App\Form\Admin;
 
-use App\Model\Product\Product;
 use Override;
 use Shopsys\FrameworkBundle\Component\Form\FormBuilderHelper;
 use Shopsys\FrameworkBundle\Form\Admin\Product\ProductFormType;
-use Shopsys\FrameworkBundle\Form\ProductsType;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -30,11 +28,7 @@ final class ProductFormTypeExtension extends AbstractTypeExtension
     #[Override]
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        /** @var \App\Model\Product\Product|null $product */
-        $product = $options['product'];
-
         $this->setSeoGroup($builder);
-        $this->setRelatedProductsGroup($builder, $product);
 
         $this->formBuilderHelper->disableFieldsByConfigurations($builder, self::DISABLED_FIELDS);
     }
@@ -47,24 +41,6 @@ final class ProductFormTypeExtension extends AbstractTypeExtension
         $builderSeoGroup = $builder->get('seoGroup');
 
         $builderSeoGroup->remove('seoH1s');
-    }
-
-    /**
-     * @param \Symfony\Component\Form\FormBuilderInterface $builder
-     * @param \App\Model\Product\Product|null $product
-     */
-    private function setRelatedProductsGroup(FormBuilderInterface $builder, ?Product $product): void
-    {
-        if (!($product !== null && $product->isVariant())) {
-            $relatedProductsGroupBuilder = $builder
-                ->create('relatedProducts', ProductsType::class, [
-                    'required' => false,
-                    'main_product' => $product,
-                    'label' => t('Related products'),
-                ]);
-
-            $builder->add($relatedProductsGroupBuilder);
-        }
     }
 
     /**
